DjangoWebProject/requirements.txt
django==1.8.7
djangorestframework
markdown
django-filter
